move `SyntaxText` into module
